requirements.txt
annotated-types==0.7.0
blis==1.0.1
boto3==1.35.62
botocore==1.35.62
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
ctransformers==0.2.27
cymem==2.0.8
dataclasses-json==0.6.7
distlib==0.3.8
faiss-cpu==1.9.0
filelock==3.16.0
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-google-genai==2.0.4
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.2
langcodes==3.4.1
langsmith==0.1.141
language_data==1.2.0
marisa-trie==1.2.1
marshmallow==3.23.1
murmurhash==1.0.10
numpy==2.1.3
orjson==3.10.11
packaging==24.2
platformdirs==4.3.6
preshed==3.0.9
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.5
sentence-transformers==3.2.1
sentencepiece==0.2.0
setuptools==75.5.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.1
tenacity==9.0.0
thinc==8.3.2
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
transformers==4.46.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.4
wasabi==1.1.3
weasel==0.4.1
wheel==0.45.0
